Allow setting the tile chart width and height
Concatenation of charts somehow sets the height or width signal to zero
in vega. In those cases, the childHeight and childWidth signals are
set instead, but we cannot use these because they don't always exists
and referencing an undefined signal throws an error.

Instead, just allow using a manually provided width or height to allow
overriding the erroneous height or width signals.