Enable concatenation of charts with tiles (#39)
* Make data layer name unique

Add a unique suffix to the tile data layer name. This ensures that a
basemap can be used in a concatenated chart, whereas this used to fail
with: Duplicate data set name: "tile_list".

* Allow setting the tile chart width and height

Concatenation of charts somehow sets the height or width signal to zero
in vega. In those cases, the childHeight and childWidth signals are
set instead, but we cannot use these because they don't always exists
and referencing an undefined signal throws an error.

Instead, just allow using a manually provided width or height to allow
overriding the erroneous height or width signals.

* Apply linter

* Fix typo

* More linting

---------

Co-authored-by: Stefan Binder <binder_stefan@outlook.com>